Clean-up v1
